feat: demais operações básicas
